requirements.txt
Django==1.10.3
django-kronos==1.0
python-crontab==2.1.1
python-dateutil==2.5.3
requests==2.11.1
six==1.10.0
